Higher quality video background for universe hero (#264)
https://www.loom.com/share/64db9edca29146259d9186f63d524bb0?sid=9dc97c0b-9e6c-45f4-add1-d8f5fcf11a24

I uploaded a high quality version of the background video to CloudFlare
so we can load it in using adaptive streaming. So the quality of the
video will adjust based on the user's internet connection.

I had to add an `hls` loader script since Safari is the only browser
that can support hls natively at this time.

I also added `transform: scale(1.01);` to the video because the source
video had a black border, this is just a hacky way to fix that without
asking for another video to be rendered, I think its fine.